# Tools/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "JongSnam.Services",
    "version": "v1"
  },
  "paths": {
...
    "/api/Store": {
      "get": {
        "tags": [
          "Store"
        ],
        "operationId": "GetStores",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "currentPage",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StoreDtoBasePagingDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Store"
        ],
        "operationId": "AddStore",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StoreRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Store/{id}": {
      "get": {
        "tags": [
          "Store"
        ],
        "operationId": "GetStoreById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StoreDetailDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Store"
        ],
        "operationId": "UpdateStore",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateStoreRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/Store/{ownerId}/Detail": {
      "get": {
        "tags": [
          "Store"
        ],
        "operationId": "GetYourStores",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ownerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "currentPage",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/YourStoreBasePagingDto"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/User": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetUsers",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserDto"
              }
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "CreateUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/User/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetUserById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserDto"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/User/{id}/ChangePassword": {
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "ChangePassword",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ChangePasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    },
    "/api/User/{id}/Update": {
      "put": {
        "tags": [
          "User"
        ],
        "operationId": "UpdateUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ProblemsDetailDto"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "ProblemsDetailDto": {
      "type": "object",
      "properties": {
        "detail": {
          "type": "string"
        },
        "extensions": {
          "type": "object",
          "additionalProperties": { },
          "readOnly": true
        },
        "instance": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "errorCode": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "UserDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "contactMobile": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "userType": {
          "type": "string"
        },
        "isLoggedIn": {
          "type": "boolean"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "StoreDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "rating": {
          "format": "double",
          "type": "number"
        },
        "officeHours": {
          "type": "string"
        },
        "isOpen": {
          "type": "boolean"
        },
        "image": {
          "type": "string"
        }
      }
    },
    "StoreDtoBasePagingDto": {
      "type": "object",
      "properties": {
        "currentPage": {
          "format": "int32",
          "type": "integer"
        },
        "totalPage": {
          "format": "int32",
          "type": "integer"
        },
        "collection": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StoreDto"
          }
        }
      }
    },
    "StoreRequest": {
      "required": [
        "address",
        "contactMobile",
        "districtId",
        "isOpen",
        "name",
        "ownerId",
        "provinceId",
        "subDistrictId"
      ],
      "type": "object",
      "properties": {
        "ownerId": {
          "format": "int32",
          "type": "integer"
        },
        "image": {
          "type": "string"
        },
        "name": {
          "format": "string",
          "maxLength": 50,
          "type": "string"
        },
        "address": {
          "format": "string",
          "maxLength": 50,
          "type": "string"
        },
        "subDistrictId": {
          "format": "int32",
          "type": "integer"
        },
        "districtId": {
          "format": "int32",
          "type": "integer"
        },
        "provinceId": {
          "format": "int32",
          "type": "integer"
        },
        "contactMobile": {
          "format": "tel",
          "maxLength": 10,
          "minLength": 9,
          "type": "string"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longtitude": {
          "format": "double",
          "type": "number"
        },
        "officeHours": {
          "maxLength": 150,
          "type": "string"
        },
        "isOpen": {
          "type": "boolean"
        },
        "rules": {
          "maxLength": 50,
          "type": "string"
        }
      }
    },
    "StoreDetailDto": {
      "type": "object",
      "properties": {
        "ownerId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "subDistrictId": {
          "format": "int32",
          "type": "integer"
        },
        "districtId": {
          "format": "int32",
          "type": "integer"
        },
        "provinceId": {
          "format": "int32",
          "type": "integer"
        },
        "contactMobile": {
          "type": "string"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longtitude": {
          "format": "double",
          "type": "number"
        },
        "rules": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "isOpen": {
          "type": "boolean"
        },
        "officeHours": {
          "type": "string"
        }
      }
    },
    "UpdateStoreRequest": {
      "required": [
        "address",
        "contactMobile",
        "district",
        "isOpen",
        "name",
        "province",
        "subDistrict"
      ],
      "type": "object",
      "properties": {
        "image": {
          "type": "string"
        },
        "name": {
          "format": "string",
          "maxLength": 50,
          "type": "string"
        },
        "address": {
          "format": "string",
          "maxLength": 50,
          "type": "string"
        },
        "subDistrict": {
          "format": "int32",
          "type": "integer"
        },
        "district": {
          "format": "int32",
          "type": "integer"
        },
        "province": {
          "format": "int32",
          "type": "integer"
        },
        "contactMobile": {
          "format": "tel",
          "maxLength": 10,
          "minLength": 9,
          "type": "string"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longtitude": {
          "format": "double",
          "type": "number"
        },
        "officeHours": {
          "maxLength": 150,
          "type": "string"
        },
        "isOpen": {
          "type": "boolean"
        },
        "rules": {
          "maxLength": 50,
          "type": "string"
        }
      }
    },
    "YourStore": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "image": {
          "type": "string"
        }
      }
    },
    "YourStoreBasePagingDto": {
      "type": "object",
      "properties": {
        "currentPage": {
          "format": "int32",
          "type": "integer"
        },
        "totalPage": {
          "format": "int32",
          "type": "integer"
        },
        "collection": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/YourStore"
          }
        }
      }
    },
    "UserRequest": {
      "required": [
        "confirmPassword",
        "contactMobile",
        "email",
        "firstName",
        "lastName",
        "password",
        "userTypeId"
      ],
      "type": "object",
      "properties": {
        "email": {
          "format": "email",
          "maxLength": 50,
          "type": "string"
        },
        "password": {
          "maxLength": 20,
          "minLength": 8,
          "type": "string"
        },
        "confirmPassword": {
          "maxLength": 20,
          "minLength": 8,
          "type": "string"
        },
        "firstName": {
          "format": "string",
          "maxLength": 50,
          "type": "string"
        },
        "lastName": {
          "format": "string",
          "maxLength": 50,
          "type": "string"
        },
        "address": {
          "format": "string",
          "maxLength": 150,
          "type": "string"
        },
        "contactMobile": {
          "format": "tel",
          "maxLength": 10,
          "minLength": 9,
          "type": "string"
        },
        "imageProfile": {
          "type": "string"
        },
        "userTypeId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "ChangePasswordRequest": {
      "required": [
        "confirmNewPassword",
        "newPassword",
        "oldPassword"
      ],
      "type": "object",
      "properties": {
        "oldPassword": {
          "maxLength": 20,
          "minLength": 8,
          "type": "string"
        },
        "newPassword": {
          "maxLength": 20,
          "minLength": 8,
          "type": "string"
        },
        "confirmNewPassword": {
          "maxLength": 20,
          "minLength": 8,
          "type": "string"
        }
      }
    },
    "UpdateUserRequest": {
      "required": [
        "contactMobile",
        "email",
        "firstName",
        "lastName"
      ],
      "type": "object",
      "properties": {
        "email": {
          "format": "email",
          "maxLength": 50,
          "type": "string"
        },
        "firstName": {
          "format": "string",
          "maxLength": 50,
          "type": "string"
        },
        "lastName": {
          "format": "string",
          "maxLength": 50,
          "type": "string"
        },
        "address": {
          "format": "string",
          "maxLength": 150,
          "type": "string"
        },
        "contactMobile": {
          "format": "tel",
          "maxLength": 10,
          "minLength": 9,
          "type": "string"
        },
        "imageProfile": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer 12345abcdef"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}